FL/cnn-img/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cnn-img"
version = "1.0.0"
description = ""
license = "Apache-2.0"
dependencies = [
    "flwr[simulation]==1.13.0",
    "flwr-datasets[vision]==0.3.0",
    "torch==2.2",
    "torchvision==0.17.1",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app]
publisher = "CNN-IMG"

[tool.flwr.app.components]
serverapp = "cnn_img.server_app:app"
clientapp = "cnn_img.client_app:app"

[tool.flwr.app.config]
num-server-rounds = 10
fraction-fit = 0.5
local-epochs = 50
server-device = "cpu"
use-wandb = true

[tool.flwr.federations]
default = "local-simulation"

[tool.flwr.federations.local-simulation]
options.num-supernodes = 100

[tool.flwr.federations.localhost-gpu]
options.num-supernodes = 50
options.backend.client-resources.num-cpus = 4 # each ClientApp assumes to use 4CPUs
options.backend.client-resources.num-gpus = 0.25 # at most 4 ClientApps will run in a given GPU
